Conflicts:
	src/cc1/Makefile.dep
	src/cc1/decl.c
	src/cc1/decl.h
	src/cc1/fold.c
	src/cc1/gen_asm.c
	src/cc1/ops/expr_op.c
	src/cc1/out/asm.c
	src/cc1/out/out.c
	src/cc1/out/out.h
	src/cc1/out/vstack.h
	src/cc1/out/x86_64.c

+++ test2/test.pl
@@ -41,7 +41,7 @@ for(@ARGV){
 $target = "./$target";
 
 END {
-	unlink $target;
+	unlink $target if defined $target;
 }
 
 my %vars = (
